requirements.txt
streamlit
pandas
duckdb
openai
python-dotenv